fix issues with tunnel testing
The addresses for the tunnels were not correctly assigned, leading
to some tunnels sharing the same local-ip/remote-ip pairs.

The tearnDown code was not removing the dummy interface created
and required for testing.

Still two tunnels are not deleted as the current code is not using
'ip tunnel delete' for ip6gre (most likely 'ip link delete' instead)

This will be resolved when the tunnel interface code is moved
to python/vyos-1x